COMMERCE-11740 autogenerated code (gw buildService)

+++ modules/apps/commerce/commerce-inventory-api/src/main/java/com/liferay/commerce/inventory/model/CommerceInventoryReplenishmentItemModel.java
@@ -11,6 +11,8 @@ import com.liferay.portal.kernel.model.MVCCModel;
 import com.liferay.portal.kernel.model.ShardedModel;
 import com.liferay.portal.kernel.model.StagedAuditedModel;
 
+import java.math.BigDecimal;
+
 import java.util.Date;
 
 import org.osgi.annotation.versioning.ProviderType;
@@ -245,14 +247,14 @@ public interface CommerceInventoryReplenishmentItemModel
 	 *
 	 * @return the quantity of this commerce inventory replenishment item
 	 */
-	public int getQuantity();
+	public BigDecimal getQuantity();
 
 	/**
 	 * Sets the quantity of this commerce inventory replenishment item.
 	 *
 	 * @param quantity the quantity of this commerce inventory replenishment item
 	 */
-	public void setQuantity(int quantity);
+	public void setQuantity(BigDecimal quantity);
 
 	/**
 	 * Returns the sku of this commerce inventory replenishment item.
