+++ modules/apps/commerce/commerce-inventory-api/src/main/java/com/liferay/commerce/inventory/service/CommerceInventoryReplenishmentItemLocalService.java
@@ -28,6 +28,7 @@ import com.liferay.portal.kernel.util.OrderByComparator;
 import java.io.Serializable;
 
 import java.math.BigDecimal;
+
 import java.util.Date;
 import java.util.List;
 
@@ -75,9 +76,9 @@ public interface CommerceInventoryReplenishmentItemLocalService
 
 	public CommerceInventoryReplenishmentItem
 			addCommerceInventoryReplenishmentItem(
-			String externalReferenceCode, long userId,
-			long commerceInventoryWarehouseId, Date availabilityDate,
-			BigDecimal quantity, String sku, String unitOfMeasureKey)
+				String externalReferenceCode, long userId,
+				long commerceInventoryWarehouseId, Date availabilityDate,
+				BigDecimal quantity, String sku, String unitOfMeasureKey)
 		throws PortalException;
 
 	/**
@@ -314,7 +315,7 @@ public interface CommerceInventoryReplenishmentItemLocalService
 	public int getCommerceInventoryReplenishmentItemsCount();
 
 	@Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
-	public long getCommerceInventoryReplenishmentItemsCount(
+	public BigDecimal getCommerceInventoryReplenishmentItemsCount(
 		long commerceInventoryWarehouseId, String sku);
 
 	@Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
@@ -366,9 +367,9 @@ public interface CommerceInventoryReplenishmentItemLocalService
 
 	public CommerceInventoryReplenishmentItem
 			updateCommerceInventoryReplenishmentItem(
-			String externalReferenceCode,
-			long commerceInventoryReplenishmentItemId,
-			Date availabilityDate, BigDecimal quantity, long mvccVersion)
+				String externalReferenceCode,
+				long commerceInventoryReplenishmentItemId,
+				Date availabilityDate, BigDecimal quantity, long mvccVersion)
 		throws PortalException;
 
 }
